Update to linux 6.4.3 and to match live machines
Adjust kernel config slightly to improve consistency between machines.

Provide explicit recipes for partitioning and making filesystems.

Creating and editing boot entries via the UEFI shell is inconvenient
and awkward. Boot anansi from the default bootloader path with kernel
command lines compiled in, eliminating the need for manually-configured
boot entries. The initramfs needs to be compiled into the kernel, so
adjust the build scripts to allow for this.

Adjust /etc/init to run agetty only on terminals listed in /proc/consoles.
Disable kernel messages on the consoles before offering a login prompt.

Update etc and images to match the live machines, which in turn have been
updated to the latest versions of userspace packages.